+++ adapters/AdColony/CHANGELOG.md
@@ -1,6 +1,12 @@
 # AdColony Mediation Adapter for Google Mobile Ads SDK for iOS
 
-## 3.0.6.0
+## Version 3.1.1.0
+- Verified compatibility with AdColony SDK 3.1.1.
+
+## Version 3.1.0.0
+- Verified compatibility with AdColony SDK 3.1.0.
+
+## Version 3.0.6.0
 - Changed the version naming system to
   [AdColony SDK version].[adapter patch version].
 - Updated the minimum required AdColony SDK to v3.0.6.